api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.11.0",
    "django==5.2.8",
    "django-ckeditor==6.7.3",
    "django-cors-headers==4.9.0",
    "django-filter==25.2",
    "django-js-asset==3.1.2",
    "djangorestframework==3.16.1",
    "pillow==12.0.0",
    "psycopg2-binary==2.9.11",
    "sqlparse==0.5.3",
    "uv==0.9.11",
]
